% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-general.R
\name{extractYFromXTable}
\alias{extractYFromXTable}
\title{extract a specific column '\code{y}' from a \code{table}
of the database 'W0004_00_Waterbirds' corresponding to 
specified '\code{xValue}' in column '\code{x}'}
\usage{
extractYFromXTable(x, y, table, ch, distinct = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{list with value for query parameter(s)
List should be named by query columns in \code{table}.}

\item{y}{name of the column(s) in \code{table} of output}

\item{table}{name of the database table to query}

\item{ch}{database connection, created with 
the \code{odbcConnect} function of the \code{RODBC} package}

\item{distinct}{logical, if TRUE (FALSE by default) extract distinct elements only}

\item{verbose}{logical, if TRUE, print the corresponding SQL query in a message}
}
\value{
query results, data.frame if multiple columns are specified with '\code{y}',
otherwise sorted vector
}
\description{
extract a specific column '\code{y}' from a \code{table}
of the database 'W0004_00_Waterbirds' corresponding to 
specified '\code{xValue}' in column '\code{x}'
}
\author{
Laure Cougnaud
}
