% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-getData.R
\name{getData}
\alias{getData}
\title{extract data of interest from a certain survey, location, taxon and survey season}
\usage{
getData(surveyId, locationId, taxonId, surveyseason, ch)
}
\arguments{
\item{surveyId}{survey identifier(s)}

\item{locationId}{location identifier(s)}

\item{taxonId}{taxon identifier(s)}

\item{surveyseason}{survey season name(s)}

\item{ch}{database connection, created with 
the \code{odbcConnect} function of the \code{RODBC} package}
}
\value{
data.frame with data of interest
}
\description{
extract data of interest from a certain survey, location, taxon and survey season
}
\author{
Laure Cougnaud
}
