% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-getLocationFromRegionLocationId.R
\name{getLocationFromRegionLocationId}
\alias{getLocationFromRegionLocationId}
\title{extract location(s) name available for a certain region and from a set of location identifiers.}
\usage{
getLocationFromRegionLocationId(regionName, locationId, ch)
}
\arguments{
\item{regionName}{name of the survey, e.g. 'Antwerpen'}

\item{locationId}{location identifier}

\item{ch}{database connection, created with 
the \code{odbcConnect} function of the \code{RODBC} package}
}
\value{
list with two elements
\itemize{
\item{'surveyId'}{survey identifier}
\item{'regioWVNaam'}{region name}
}vector of regions available for specified surveys
}
\description{
Location identifiers should be provided to restrict to locations
sampled in a certain survey.
}
\author{
Laure Cougnaud
}
