% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPathReport.R
\name{getPathStartTemplate}
\alias{getPathStartTemplate}
\title{get path of start template available in the package}
\usage{
getPathStartTemplate()
}
\value{
string with path of start template
}
\description{
get path of start template available in the package
}
\author{
Laure Cougnaud and Kirsten Van Hoorde
}
