% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-getPersonKey.R
\name{getPersonKey}
\alias{getPersonKey}
\title{get key of person stored in database from email (provided in ShinyProxy)}
\usage{
getPersonKey(ch, email)
}
\arguments{
\item{ch}{database connection, created with 
the \code{odbcConnect} function of the \code{RODBC} package}

\item{email}{string with email}
}
\value{
string with person key
}
\description{
get key of person stored in database from email (provided in ShinyProxy)
}
\author{
Laure Cougnaud
}
