% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-general.R
\name{getQuery}
\alias{getQuery}
\title{create a SQL query}
\usage{
getQuery(x = NULL, y, table, distinct = FALSE)
}
\arguments{
\item{x}{list with value for query parameter(s)
List should be named by query columns in \code{table}.}

\item{y}{name of the column(s) in \code{table} of output}

\item{table}{name of the database table to query}

\item{distinct}{logical, if TRUE (FALSE by default) extract distinct elements only}
}
\value{
string with SQL query
}
\description{
create a SQL query
}
\author{
Laure Cougnaud
}
