% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-general.R
\name{getQueryJoin}
\alias{getQueryJoin}
\title{build join part of SQL query}
\usage{
getQueryJoin(tableX, tableY, key, typeJoin = c("none", "left", "right"))
}
\arguments{
\item{tableX}{name of first SQL table to join from}

\item{tableY}{name of second SQL table to join to}

\item{key}{key used to join the two tables}

\item{typeJoin}{string with type of join: 'none', 'left' or 'right'}
}
\value{
string with join part of SQL query
}
\description{
build join part of SQL query
}
\author{
Laure Cougnaud
}
