% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-getRegioFromSurvey.R
\name{getRegioFromSurvey}
\alias{getRegioFromSurvey}
\title{extract region(s) available for a certain survey}
\usage{
getRegioFromSurvey(surveyName, ch)
}
\arguments{
\item{surveyName}{name of the survey, e.g. 'Slaapplaatstellingen Aalscholvers'}

\item{ch}{database connection, created with 
the \code{odbcConnect} function of the \code{RODBC} package}
}
\value{
list with two elements
\itemize{
\item{'surveyId': }{survey identifier}
\item{'location': }{identifiers of available location(s for the survey}
\item{'regioName': }{name of available region(s) for the survey}
}vector of regions available for specified surveys
}
\description{
extract region(s) available for a certain survey
}
\author{
Laure Cougnaud
}
