% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-getSeasonFromSurveyLocationTaxon.R
\name{getSeasonFromSurveyLocationTaxon}
\alias{getSeasonFromSurveyLocationTaxon}
\title{extract season(s) available for a survey, location and taxon}
\usage{
getSeasonFromSurveyLocationTaxon(surveyId, locationId, taxon, ch)
}
\arguments{
\item{surveyId}{survey identifier}

\item{locationId}{location identifier}

\item{taxon}{taxon name}

\item{ch}{ch database connection, created with 
the \code{odbcConnect} function of the \code{RODBC} package}
}
\value{
vector with available seasons
}
\description{
extract season(s) available for a survey, location and taxon
}
\author{
Laure Cougnaud
}
