% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-getTaxonFromTaxonGroupTaxonId.R
\name{getTaxonFromTaxonGroupTaxonId}
\alias{getTaxonFromTaxonGroupTaxonId}
\title{extract taxon(s) available for a taxon group and from set of taxon identifiers}
\usage{
getTaxonFromTaxonGroupTaxonId(taxonId, taxonGroup, ch)
}
\arguments{
\item{taxonId}{taxon identifier(s)}

\item{taxonGroup}{taxon group(s)}

\item{ch}{database connection, created with 
the \code{odbcConnect} function of the \code{RODBC} package}
}
\value{
vector with available taxons (common names)
}
\description{
Taxon identifiers should be provided to restrict to taxons
sampled in a certain survey.
}
\author{
Laure Cougnaud
}
