% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-getTaxonGroupFromLocationSurvey.R
\name{getTaxonGroupFromLocationSurvey}
\alias{getTaxonGroupFromLocationSurvey}
\title{extract taxon group(s) available for a certain survey location and survey}
\usage{
getTaxonGroupFromLocationSurvey(locationName, surveyId, personKey, ch)
}
\arguments{
\item{locationName}{name of the location, e.g. 'Galgenschoor (Schelde Lillo-Fort - Containerkaai) (RO)'}

\item{surveyId}{identifier of the survey, e.g. 6}

\item{personKey}{integer with person key}

\item{ch}{database connection, created with 
the \code{odbcConnect} function of the \code{RODBC} package}
}
\value{
list with two elements
\itemize{
\item{'locationId': }{location identifier}
\item{'taxonId': }{identifier of available taxon(s)}
\item{'taxonGroup': }{name of available taxon group(s) for the survey}
}
}
\description{
extract taxon group(s) available for a certain survey location and survey
}
\author{
Laure Cougnaud
}
