% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-getTeldatumSeason.R
\name{getTeldatumSeason}
\alias{getTeldatumSeason}
\title{get data.frame with all possible sampling date for each season}
\usage{
getTeldatumSeason(ch)
}
\arguments{
\item{ch}{database connection, created with 
the \code{odbcConnect} function of the \code{RODBC} package}
}
\value{
data.frame with columns 'surveyseason' and 'teldatum'
}
\description{
get data.frame with all possible sampling date for each season
}
\author{
Laure Cougnaud
}
