% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCountsDistributions.R
\name{plotCountsDistribution}
\alias{plotCountsDistribution}
\title{generate a plot with counts distribution}
\usage{
plotCountsDistribution(inputDataPlot, teldatumSeasonDf, typePlot = c("ggplot",
  "plotly"))
}
\arguments{
\item{inputDataPlot}{data.frame with input data}

\item{teldatumSeasonDf}{data.frame with available teldatum per surveyseason}

\item{typePlot}{string, 'ggplot' for static plot of 'plotly' for interactive plot}
}
\value{
a code{ggplot} or a \code{plotly} object, depending on the parameter \code{typePlot}
}
\description{
generate a plot with counts distribution
}
\author{
Laure Cougnaud
}
