% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readData}
\alias{readData}
\title{simple wrapper to extract all data tables from set of files}
\usage{
readData(basePath)
}
\arguments{
\item{basePath}{base path where all the tables are present in csv format}
}
\value{
list of data.frame (each one is a different table)
}
\description{
simple wrapper to extract all data tables from set of files
}
\author{
Laure Cougnaud
}
