% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runWatervogels.R
\name{runWatervogels}
\alias{runWatervogels}
\title{launch watervogels shiny application}
\usage{
runWatervogels()
}
\value{
no returned value, the shiny application is launched
}
\description{
launch watervogels shiny application
}
\author{
Laure Cougnaud
}
