% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-getProjects.R
\name{getProjects}
\alias{getProjects}
\title{get available projects}
\usage{
getProjects(ch)
}
\arguments{
\item{ch}{database connection, created with 
the \code{odbcConnect} function of the \code{RODBC} package}
}
\value{
vector with available projects
}
\description{
get available projects
}
\author{
Laure Cougnaud
}
